% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amexmetric.R
\name{amex_metric}
\alias{amex_metric}
\title{Compute the metric for the Amex competition}
\usage{
amex_metric(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
Compute the metric for the Amex competition
}
